+++ src/components/common/AccountBtn.module.scss
@@ -1,0 +1,72 @@
+@import '@/_variables.scss';
+
+.container {
+  width: 150px;
+  height: 150px;
+  background-color: $bg-sub-grey;
+  border-radius: 18px;
+  padding: 10px;
+  margin-top: 10px;
+  .bank {
+    height: 100%;
+    display: flex;
+    position: relative;
+    flex-direction: column;
+    align-items: start;
+    justify-content: space-between;
+    // background-color: green;
+    .bankLogo {
+      position: absolute;
+      top: 5px;
+      right: 5px;
+      width: 60px;
+      height: 60px;
+      background-color: $bg-sub-grey;
+      mix-blend-mode: multiply;
+      z-index: 1;
+      border-radius: 50%;
+    }
+    .bankInfo {
+      position: absolute;
+      top: 65px;
+      left: 10px;
+      width: 100%;
+      height: 40%;
+      display: flex;
+      flex-direction: column;
+      justify-content: space-between;
+      // background-color: yellow;
+      .bankName {
+        width: 90%;
+        text-align: left;
+        font-size: 1.1rem;
+        font-weight: 700;
+        margin-bottom: 2px;
+      }
+      .changes {
+        position: absolute;
+        top: 20px;
+        text-align: right;
+        width: 85%;
+        height: 70%;
+        padding: 5px;
+        // border-radius: 10px;
+        // background-color: $bg-sub-grey;
+        // mix-blend-mode: multiply;
+        .title {
+          font-weight: 800;
+          color: $text-sub-grey;
+          margin: 3px;
+        }
+        .change {
+          font-size: 1.1rem;
+          font-weight: 800;
+          margin: 3px;
+        }
+      }
+    }
+  }
+  &:active {
+    border: 1px solid $bg-primary-grey;
+  }
+}
