+++ src/components/common/BankBtn.module.scss
@@ -1,0 +1,47 @@
+@import '@/_variables.scss';
+
+.container {
+  width: 200px;
+  height: 100px;
+  background-color: $bg-sub-grey;
+  border-radius: 18px;
+  padding: 10px;
+  &:active {
+    border: 1px solid $bg-primary-grey;
+  }
+  .bank {
+    height: 100%;
+    display: flex;
+    position: relative;
+    align-items: center;
+    justify-content: space-betweec;
+    // background-color: red;
+    .bankLogo {
+      position: absolute;
+      left: 10px;
+      width: 70px;
+      height: 70px;
+      background-color: $bg-sub-grey;
+      mix-blend-mode: multiply;
+      z-index: 1;
+      border-radius: 50%;
+    }
+    .bankInfo {
+      position: absolute;
+      // top: 65px;
+      right: 13px;
+      width: 100%;
+      height: 45%;
+      display: flex;
+      flex-direction: column;
+      justify-content: center;
+      align-items: end;
+      text-align: right;
+      // background-color: yellow;
+      .bankName {
+        font-size: 1.1rem;
+        font-weight: 800;
+      }
+    }
+  }
+}
