--- src/components/mypage/bank/LinkedBank.module.scss
@@ -1,64 +1,0 @@
-@import '@/_variables.scss';
-
-.container {
-  box-sizing: border-box;
-  width: 150px;
-  height: 150px;
-  border: none;
-  background-color: $bg-primary-white;
-  border-radius: 20px;
-  padding: 10px;
-  margin: 10px;
-  box-shadow: 1px 1px 2px $point-orange;
-  cursor: pointer;
-
-  .bank {
-    height: 100%;
-    display: flex;
-    justify-content: space-between;
-    .bankLogo {
-      width: 40px;
-      height: 40px;
-      margin-right: 10px;
-    }
-    .bankInfo {
-      display: flex;
-      flex-direction: column;
-      justify-content: space-between;
-      flex-wrap: nowrap;
-      text-align: right;
-      .bankName {
-        margin-top: 10px;
-        font-size: 16px;
-        font-weight: 700;
-      }
-      .accountNumber {
-        font-size: 12px;
-        word-wrap: break-word;
-        color: $text-sub-grey;
-      }
-      .changes {
-        .title {
-          margin-bottom: 5px;
-          font-weight: 700;
-          color: $text-sub-grey;
-        }
-        .change {
-          margin-bottom: 10px;
-          font-size: 18px;
-          font-weight: 700;
-        }
-      }
-    }
-  }
-}
-.container:hover {
-  position: relative;
-  top: -2px;
-}
-
-.active {
-  box-shadow: 3px 3px 3px royalblue;
-  position: relative;
-  top: 4px;
-}
