+++ src/components/mypage/bank/SelectAccount.module.scss
@@ -1,0 +1,9 @@
+@import '@/_variables.scss';
+
+.btn {
+  border: none;
+  background-color: transparent;
+  position: relative;
+  cursor: pointer;
+
+}
