+++ src/components/mypage/bank/SelectBank.module.scss
@@ -1,0 +1,10 @@
+@import '@/_variables.scss';
+
+.btn {
+  position: relative;;
+  border: none;
+  background-color: transparent;
+  margin: 5px 0;
+  cursor: pointer;
+
+}
