+++ src/components/mypage/productList/JjimItem.module.scss
@@ -52,7 +52,7 @@
   }
   .itemPrice {
     position: relative;
-    left: -33px;
+    left: -30px;
     font-weight: 800;
   }
 }
