+++ src/pages/mypage/account/addAccount/index.module.scss
@@ -1,7 +1,41 @@
 @import 'src/_variables.scss';
 
 .container {
-  width: 1440px;
-  height: 500px;
-  background-color: rgb(205, 180, 78);
+  margin: auto;
+  width: $mypage-width;
+  height: $mypage-height;
+  background-color: $bg-sub-grey;
+  display: flex;
+  margin-bottom: 30px;
+  .section {
+    position: relative;
+    width: 65%;
+    height: 80%;
+    margin: auto;
+    mix-blend-mode: multiply;
+    // background-color: rgba(0, 128, 0, 0.354);
+    .title {
+      font-size: 1.4rem;
+      letter-spacing: -0.5;
+      font-weight: 800;
+    }
+    .line {
+      margin: 20px auto 25px;
+      border: 1px solid $bg-sub-grey;
+    }
+    .text {
+      font-size: 1.1rem;
+      letter-spacing: -0.5;
+      font-weight: 300;
+    }
+    .banks {
+      position: absolute;
+      top: 115px;
+      display: flex;
+      flex-wrap: wrap;
+      justify-content: space-between;
+      text-align: left;
+      font-size: 1.1rem;
+    }
+  }
 }
