+++ src/pages/mypage/account/index.module.scss
@@ -8,21 +8,92 @@
   display: flex;
   margin-bottom: 30px;
   .bankContainer {
-    width: 100%;
-    background-color: green;
+    position: relative;
+    width: 70%;
+    height: 90%;
+    margin: auto;
+    mix-blend-mode: multiply;
+    // background-color: green;
     .totallAccount {
-      margin: 30px auto;
-      width: 90%;
-      height: 35%;
-      background-color: yellow;
+      position: relative;
+      top: 0;
+      margin: auto;
+      width: 100%;
+      height: 30%;
+      border-radius: 15px;
+      background-color: $bg-sub-grey;
+      .totale {
+        position: absolute;
+        top: 30px;
+        left: 50px;
+        text-align: left;
+        font-size: 1.1rem;
+      }
+      .totaleAmount {
+        position: absolute;
+        bottom: 40px;
+        right: 50px;
+        text-align: right;
+        font-size: 2.4rem;
+        font-weight: 800;
+      }
+      .underLine {
+        position: absolute;
+        display: block;
+        height: 50px;
+        width: 300px;
+        bottom: 30px;
+        right: 50px;
+        text-align: left;
+        color: $bg-primary-grey;
+        border-bottom: 2px solid $bg-primary-grey;
+        font-size: 3rem;
+        font-weight: 800;
+        z-index: 1;
+      }
     }
     .bankAccount {
-      margin: 30px auto;
-      width: 90%;
-      height: 50%;
-      display: flex;
-      flex-wrap: wrap;
-      background-color: yellow;
+      position: absolute;
+      bottom: 0;
+      left: 2.5%;
+      margin: auto;
+      width: 95%;
+      height: 65%;
+      border-radius: 15px;
+      // background-color: green;
+      .banks {
+        width: 100%;
+        flex-wrap: wrap;
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+      }
+      .addBank {
+        width: 155px;
+        height: 155px;
+        display: flex;
+        padding: 5px;
+        margin: 10px 5px 0 3px;
+        justify-content: center;
+        align-items: center;
+        border: 3px dashed $bg-sub-grey;
+        border-radius: 18px;
+        text-decoration: none;
+        .btn {
+          position: sticky;
+          display: flex;
+          flex-direction: column;
+          justify-content: center;
+          align-items: center;
+          font-size: 7rem;
+          font-weight: 100;
+          color: $bg-primary-grey;
+          margin-bottom: 20px;
+          .add {
+            font-size: 0.9rem;
+          }
+        }
+      }
     }
   }
 }
