+++ src/pages/mypage/certProfile/editProfile/index.module.scss
@@ -22,41 +22,40 @@
     border-radius: 10px;
     background-color: $bg-primary-white;
     .content {
-      position: relative;
+      margin: auto;
       height: 80%;
       width: 80%;
       display: flex;
       flex-direction: column;
       justify-content: space-between;
       align-items: start;
-      font-size: 20px;
+      font-size: 1.1rem;
       font-weight: 300;
       text-align: left;
+      // background-color: yellowgreen;
       .innerTop {
         width: 100%;
-        margin: 3px;
+        // background-color: green;
         .title {
           font-size: 1.4rem;
           letter-spacing: -0.5;
           font-weight: 800;
         }
         .line {
-          margin: 20px auto;
+          margin: 25px auto;
         }
       }
       .innerBottom {
-        position: relative;
-        height: 100%;
-        width: 100%;
-        top: 0;
+        height: 80%;
         // background-color: green;
         .lists {
-          height: 100%;
+          height: 95%;
           display: flex;
           flex-direction: column;
           justify-content: space-between;
           align-items: center;
           margin: 0 auto;
+          // background-color: greenyellow;
           .list {
             width: 100%;
             display: flex;
@@ -65,7 +64,7 @@
             line-height: 1.3;
             .label {
               font-size: 1.1rem;
-              letter-spacing: -0.5;
+              letter-spacing: -.5;
               font-weight: 800;
               margin-right: 30px;
               width: 100px;
@@ -82,7 +81,7 @@
                 padding: 15px 20px;
                 border-radius: 5px;
                 border: solid 1px $bg-primary-grey;;
-                font-size: 1rem;
+                font-size: .9rem;
                 &:focus {
                   border: solid 2px $point-red;
                   caret-color: $point-red;
