+++ src/pages/mypage/index.tsx
@@ -43,8 +43,11 @@ export default function MyPage() {
                 <li className={`${styles.listTitle} ${styles.Items}`}>
                   {visible ? '찜한 상품' : '구매 내역'}
                   <div className={styles.listText}>
-                    <button className={`${styles.aTag} ${styles.btn} ${styles.listLink}`} onClick={handleVisible}>
-                      {visible ? '❤️ 찜한 상품' : '💰 구매 내역'}
+                    <button
+                      className={`${styles.aTag} ${styles.btn} ${styles.listLink} ${styles.visible}`}
+                      onClick={handleVisible}
+                    >
+                      {visible ? '❤️ 찜한 상품 ' : '💰 구매 내역 '}
                     </button>
                   </div>
                 </li>
@@ -79,7 +82,7 @@ export default function MyPage() {
                   {visible ? '찜한 상품' : '구매 내역'}
                   <div className={styles.listText}>
                     <button className={`${styles.aTag} ${styles.btn} ${styles.listLink}`} onClick={handleVisible}>
-                      {visible ? '❤️ 찜한 상품' : '💰 구매 내역'}
+                      {visible ? '❤️ 찜한 상품 ' : '💰 구매 내역 '}
                     </button>
                   </div>
                 </li>
